Changed getStaticData to push code into indexers for less code duplication
