+++ back-end/src/main/java/index/PsqlCubeSpatialIndexer.java
@@ -8,7 +8,6 @@ import project.Canvas;
 import project.Layer;
 import project.Transform;
 import project.Project;
-import server.JsonWriter;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -214,35 +213,18 @@ public class PsqlCubeSpatialIndexer extends Indexer {
     }
 
     @Override
-    public ArrayList<ArrayList<ArrayList<String>>> getStaticData(Canvas c, ArrayList<String> predicates)
-            throws SQLException, ClassNotFoundException {
-    
-          // container for data
-          ArrayList<ArrayList<ArrayList<String>>> data = new ArrayList<>();
+    public String getStaticDataQuery(Canvas c, Layer l, int i, ArrayList<String> predicates) {
 
-          // loop over layers
-          for (int i = 0; i < c.getLayers().size(); i ++) {
-  
-              // add an empty placeholder for static layers
-              if (! c.getLayers().get(i).isStatic()) {
-                  data.add(new ArrayList<>());
-                  continue;
-              }
-  
-              // get column list string
-              String colListStr = c.getLayers().get(i).getTransform().getColStr("");
-  
-              // construct range query
-              String sql = "select " + colListStr + " from bbox_" + Config.projectName;
-              if (predicates.get(i).length() > 0)
-                  sql += " where " + predicates.get(i);
-              sql += ";";
-  
-              // run query, add to response
-              data.add(DbConnector.getQueryResult(Config.databaseName, sql));
-          }
-  
-          return data;
+        // get column list string
+        String colListStr = l.getTransform().getColStr("");
+
+        // construct range query
+        String sql = "select " + colListStr + " from bbox_" + Config.projectName;
+        if (predicates.get(i).length() > 0)
+            sql += " where " + predicates.get(i);
+        sql += ";";
+
+        return sql;
 
     }
